refactor: number/integer 필드 통합 + 열 서식 도구모음 (FormatToolbar) (#296)
number와 integer를 단일 number 타입으로 통합하고 decimal_places 옵션으로
정수/소수 구분. SpreadsheetView에 열 단위 서식 변경 도구모음 추가.

Backend:
- IsNumeric() 메서드, ExtractDecimalPlaces/EffectiveDecimalPlaces 헬퍼
- FieldInteger DDL → NUMERIC 통합, compat 매트릭스 convAlways
- api_validate/csv/dynamic 핸들러 통합 (decimal_places 기반 검증)
- seed 데이터 integer→number+decimal_places:0

Frontend:
- NumberFieldOptions에 decimal_places 추가
- isNumericField/getDecimalPlaces 헬퍼
- FieldProperties: 숫자/정수 변형 토글 → 소수자릿수 Select
- formatCell: decimal_places 기반 포맷
- FormatToolbar: 선택 열의 표시형식/소수자릿수/통화코드 변경
- DataTable onActiveCellChange → SpreadsheetView → FormatToolbar 연동
- useUpdateField mutation 추가

+++ backend/internal/migration/ddl.go
@@ -34,7 +34,7 @@ func FieldTypeToPG(ft schema.FieldType) string {
 	case schema.FieldNumber:
 		return "NUMERIC"
 	case schema.FieldInteger:
-		return "INTEGER"
+		return "NUMERIC" // unified with FieldNumber; existing INTEGER columns are left as-is
 	case schema.FieldBoolean:
 		return "BOOLEAN"
 	case schema.FieldDate:
